adding fanout

+++ cpp/include/rapidsmpf/streaming/core/fanout.hpp
@@ -14,7 +14,7 @@
  /**
   * @brief Fanout policy controlling how messages are propagated.
   */
- enum class FanoutPolicy : int {
+ enum class FanoutPolicy : uint8_t {
      /**
       * @brief Process messages as they arrive and immediately forward them.
       *
